requirements.txt
Flask==2.2.5
numpy==1.26.3
statistics
scikit-learn==1.2.1
SQLAlchemy==1.4.31
Flask-WTF==0.15.1
passlib==1.7.4
python-dotenv==0.19.2
firebase-admin==5.0.2
click==8.1.7
gunicorn==22.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==2.2.2
itsdangerous==2.1.2
WTForms==2.3.3
email-validator==1.1.3
pytz==2021.3
bcrypt
google-cloud-firestore==2.1.0
pandas
openpyxl
scipy
joblib
threadpoolctl
EntropyHub
praat-parselmouth
librosa
soundfile
matplotlib
Flask-Limiter
redis
rq